Library

Synchronous Programming

A stream function is introduced by the macro @node. Inside a node, the macro @init can be used to declare a variable as a memory. Another macro @prev can then be used to access the value of a memory variable at the previous time step.

OnlineSampling.@node — Macro

Introduced a stream function. E.g.,

@node function one()
    return 1
end

OnlineSampling.@init — Macro

Declare a variable as a memory with a default value. Can only be used inside a @node declaration.

OnlineSampling.@prev — Macro

Access the value of a memory variable at the previous time step. Can only be used inside a @node declaration.

OnlineSampling.@nodecall — Macro

Function call for stream functions.

E.g.,

@node function one()
    return 1
end

@node function two()
    x = @nodecall one()
    return x + 1
end

OnlineSampling.cst — Function

Convenience alias for Iterators.repeated

Probabilistic Programming

In a probabilistic model, random variables are introduced by rand and can be conditioned on concrete value using @observe.

OnlineSampling.@observe — Macro

Condition the model with the assumption that a random variable introduced by rand takes a concrete value.

x = rand(Normal(0, 1))     # x_t ~ N(0, 1)
@observe(x, 1.5)           # assume x = 0.5

Runtime

Stream functions – probabilistic or not – can then be executed as Julia iterators using the following macros.

OnlineSampling.@nodeiter — Macro

Turn a stream function into a julia iterator. Arguments:

T is the number of iterations (optional, default=nothing)

algo is the probabilistic runtime (optional, see Algorithms)

particles is the number of particles for the probabilistic runtime (optional, default=0)

rt is the resampling threshold (optional, default=0.5)

E.g.,

for x in @nodeiter T = 10 f() # for 10 iterations of f
    println(x)                # print the current value
end

OnlineSampling.@noderun — Macro

Unfold an iterator and return its last value Arguments:

T is the number of iterations (optional, default=nothing)

algo is the probabilistic runtime (optional, see Algorithms)

particles is the number of particles for the probabilistic runtime (optional, default=0)

rt is the resampling threshold (optional, default=0.5)

E.g.,

res = @noderun T = 10 f()

OnlineSampling.Algorithms — Type

Enum type to choose the inference algorithm in @noderun and @nodeiter. Can be one of:

particle_filter

delayed_sampling

belief_propagation

streaming_belief_propagation